requirements.txt
fastapi
uvicorn[standard]
sqlmodel
SQLAlchemy
pydantic
passlib[bcrypt]
python-jose
scikit-learn
pandas
pytest
httpx
python-multipart

comm
ipython
matplotlib
seaborn
tornado
traitlets
pydantic[email]
